+++ changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.8.10] - 2024-06-18
+
+### New
+
+- Truncate option
+
+### Fixed
+
+- More date formats for postgres being used
+
 ## [0.8.8] - 2023-07-30
 
 ### Changed